Change ndeletable in _hash_vacuum_one_page() from int to uint16
deletable in _hash_vacuum_one_page() is assigned to xl_hash_vacuum_one_page.ntuples
which has been changed to uint16 in a previous commit.

This commit ensures that the assigned value has the same type.

Doing so produces those cascading changes:

_hash_vacuum_one_page
	index_compute_xid_horizon_for_tuples
		gistprunepage
			PageIndexMultiDelete
			gistXLogDelete
	PageIndexMultiDelete
		spgRedoVacuumRedirect
		vacuumRedirectAndPlaceholder
		spgPageIndexMultiDelete
			moveLeafs
			doPickSplit
		_bt_delitems_vacuum
			btvacuumpage
		_bt_delitems_delete
			_bt_delitems_delete_check
		hash_xlog_move_page_contents
		gistvacuumpage
			gistXLogUpdate
				gistplacetopage
		hashbucketcleanup

which also fixed others missmatch on the way.

Also in some places the types have been changed from OffsetNumber to uint16.
Even if at the time of this commit the OffsetNumber is defined as uint16, it's
better to match the functions arguments types they are linked to.