Report the total amount of time that vacuum has been delayed due to cost delay
This commit adds one column: time_delayed to the pg_stat_progress_vacuum system
view to show the total amount of time in milliseconds that vacuum has been
delayed.

This uses the new parallel message type for progress reporting added
by f1889729dd.

In case of parallel worker, to avoid the leader to be interrupted too frequently
(while it might be sleeping for cost delay), the report is done only if the last
report has been done more than 1 second ago.

Having a time based approach only to throttle the reporting of the parallel
workers sounds reasonable: indeed, the more parallel workers is used, the less
the impact of the leader on the vacuum index phase duration/workload is.

Bump catversion because this changes the definition of pg_stat_progress_vacuum.