Add XIDOID in [de]construct_array_builtin()
Using construct_array_builtin() for XIDOID is a new use case that is coming with
a new module (not added in the code tree yet).

d746021de1 (which introduced construct_array_builtin()) did not take care of
XIDOID because there were no use case at that time. Now that there is one, let's
add XIDOID.