Add NB_WAIT_EVENTS in wait_event_types.h
Adding NB_WAIT_EVENTS computation in generate-wait_event_types.pl and
define its result in wait_event_types.h. That way we can get rid of the
extra row (full of NULLs) added in waitEventData[] and that could help new
code (if any) that would need to get the number of wait events.