switching PGSTAT_KIND_WAIT_EVENT to variable sized
It might be better for PGSTAT_KIND_WAIT_EVENT to be a variable sized stats kind.
That way:

* It would be easier to add custom wait events if we want to
* It would be possible to add a new wait events without having to change the
stats file format

It uses the uint32 as the hash key while the hash key is defined as uint64: that
should not be an issue but this patch does explicit casting though. That said it
might be better to use all the 64-bit (means not have the half full of zeroes) for
the hash key (better hashing distribution?) then we could imagine using something
like:

((uint64) wait_event_info) | (((uint64) wait_event_info) << 32)

for the hash key.