Add pgstat_report_activity() to IO workers
pgaio_io_perform_synchronously() correctly sets wait events while performing
IO but the IO workers are not in an active state. So, monitoring queries
filtering on "state = 'active'" in pg_stat_activity would miss the fact that
the IO workers are performing an IO (and waiting on it).

The only place to add pgstat_report_activity() is in IoWorkerMain(). The other
places where pgaio_io_perform_synchronously() is executed (pgaio_worker_submit_internal()
and pgaio_io_stage() are for non IO workers and have their state correctly set
to active already).