Create lwlocktranchelist.h
This commit extracts the predefined LWLocks tranches from lwlock.c and puts them
in a new file (namely lwlocktranchelist.h).

This way, we can include this file in lwlock.c using the same macro pattern
as we do for lwlocklist.h.

This gives us the chance to cross-check with wait_event_names.txt in
generate-lwlocknames.pl in a following commit.