Cross-check lists of predefined tranches.
Both builtin_lwlocktranchelist.h and wait_event_names.txt contain a list of all
the predefined tranches. It is easy to miss one or the other (see a493e741d32
,08b9b9e043b and c3623703f36).

This commit adds a cross-check of these lists to the script that already does
the same cross-check for lwlocklist.h.

If the lists do not match exactly, building will fail (same as in 5b1b9bce844).