Fix some 035_standby_logical_decoding.pl tests and wordsmithing
usleep is used but not imported (oversight in commit 48796a98d5ae).

"replication" is missing in two checks leading to tests that could never fail.

Also using "back" in a comment gives the wrong impression that it has been already
set that way (which is not the case on purpose).

Reported-by: Melanie Plageman <melanieplageman@gmail.com>