Get rid of pgstat_count_backend_io_op*() functions
This commit removes the functions that are incrementing the backend IO stats.
Instead, it now copies the IO pending stats to the backend IO pending stats when
the pending IO stats are flushed. This behaves the same way as for some relation
and database stats.

It's done that way to avoid incrementing the "same" counters twice as it produces
increased overhead in profiles.

Note that per-backend statistics have to be last in pgstat_report_stat() as some
of their pending stats are populated during other stats kinds flush.

Reported-by: Andres Freund <andres@anarazel.de>