Add stats_reset to pg_stat_all_tables|indexes and related views
It's possible to call pg_stat_reset_single_table_counters() on a relation but
the reset time is not reported in the related views. Adding stats_reset to
the related views for consistency with other stat kinds that also report reset
times.

This new field is not included into the pg_stat_xact_* views because the 0 values
in transaction-local stats have nothing to do with reset operations.

XXX: Bumps catversion
XXX: Bumps PGSTAT_FILE_FORMAT_ID