Don't use initial_* in InvalidatePossiblyObsoleteSlot()
818fefd8fd4 fixed race leading to incorrect conflict cause in
InvalidatePossiblyObsoleteSlot() for all possible conflicts.

However the pre 818fefd8fd4 behavior was intentional: it's possible for a slot
to advance its restart_lsn or xmin values sufficiently between when we release
the mutex and when we recheck, moving from a conflicting state to a
non conflicting state. This is safe: if the slot has caught up while we're
busy here, the resources we were concerned about (WAL segments or tuples) have
not yet been removed, and there's no reason to invalidate the slot.

This commit restores the pre 818fefd8fd4 behavior and improves a related comment
in passing. The tests should not fail sporadically on slow machines since
105b2cb3361 is in place.

Reported-by: suyu.cmj <mengjuan.cmj@alibaba-inc.com>
Reviewed-by: Masahiko Sawada <sawada.mshk@gmail.com>
Reviewed-by: Amit Kapila <amit.kapila16@gmail.com>
Discussion: https://postgr.es/m/f492465f-657e-49af-8317-987460cb68b0.mengjuan.cmj@alibaba-inc.com