Safer hash table initialization macro
Currently to create a hash table, we do things like:

ctl.keysize = sizeof(<Type>);
ctl.entrysize = sizeof(<Struct>);

but that lead to 2 possible issues:

- we manually specify the type for keysize, which could become incorrect (from
the start) or if the key member's type changes.

- it may be possible to remove the key member without the compiler noticing it.

To fix those, this commit introduces a new HASH_ELEM_INIT macro that:

- requests the key member name
- ensures that it is at offset 0
- computes the key size based on the member

so that:

- the key member is explicitly referenced in the code (preventing "unused"
false positives)
- the key size is automatically computed from the actual member type (preventing
type mismatches)
- we enforce that the key is at offset 0

Also note that this commit adds HASH_ELEM_INIT_FULL for the rare cases where
the whole struct is the key.